updates to boundaries
added ending to map. THIS SHOULD BE WHERE WE PRACTICE WHAT HAPPENS WHEN
WE BEAT A LEVEL.

If anyone wants to implement a level completion, I'm thinking basically
create an invisible sprite, place it at the right coordinates so the
player collides with it at the end, then when they collide, do something
so the game knows you won.