Added ending to level, bug fixes to enemies and shooting physics
Added an ending to the level and fixed some bugs with the shooting and
enemy related physics to improve the gameplay.

- When hero reaches the end of the level, he 'goes into' a spaceship and
the enemy chaser explodes. This sequence can certainly be improved tho
but it's a start.
- Victory text shows when level is completed.
- Enemies can only be shot when on the screen.
- Fixed enemy sprite to run flat on ground
- If hero shoots enemy, then the enemy's fired bullet is 'killed' as
well.
- Enemies no longer still shoot after they're killed.

The next thing that has to be done is the level 'resetting' when the
hero dies. This is a must by next Friday.