Added AI for running enemies
Added some reactionary AI for the running enemies. Based on the hero's
movement the enemies will (sort of) act accordingly. Still some
debugging to be done on the flipping motion so someone could work on
that. Also shrank the enemy sprites a bit.

Since these enemies have AI now we can decide as a team if we want to
keep the enemy chaser and add more functionality to it or remove it
altogether.