Big Ol' Commit. Please read Description
- new Enemy AI. The new enemy AI is mainly in Enemy.ts
# The enemy will always try to match your gravity within a certain
range, but it activates a cooldown. This cooldown is necessary to give
Hero the advantage.
-Enemy now contains a reference to hero (agent oriented programming) and
level.
-Both levels 1 and 2  have Enemy Spawn points. They are beatable (there
are occasional parts where enemy spawns fuck you but that's part of the
game) but they can be difficult.
- Removed BossLevel from level progression as it currently seems
unbeatable.
-added sounds
-Updated "Level_3.json" - I hope no one minds!! I just added and moved
some things (currently Phaser doesn't support ramped tiles). if anyone
wants to try and implement this level feel free.