Level 1 = noob level.
Level 1 is now a (tentative) level called LevelNoob. if we like it, it
stays, but we'll need more maps.
Changed death text to stuff that's more polite. Chris might not have
appreciated being told he sucks and is a loser by students.
also added and included the audio file 'collision' that was not
implemented in the 'Audio Added' update correctly