Additional PowerUps
Added PowerUps that currently add to the total Score of the game.

* Star - Adds 1000 to total score

* Diamond - Adds 2000 to total score

* Clock - Currently adds 1000 to total score, but working on changing
the game speed upon impact (i.e. slow-mo or speed-up) // IN PROGRESS,
NOT COMPLETE
* Key - Currently adds 1000 to total score, but want to change it so
that the hero goes straight to the next level upon impact, (i.e. "key"
to the next level) // IN PROGRESS, NOT COMPLETE

* Magic/Orb - Currently adds 1000 to total score, but want to change it
so that it does something cool...haven't thought of something creative
enough yet...  // IN PROGRESS, NOT COMPLETE