Invincibility PowerUp
When Samus collides with the orb/Invincibility PowerUp, the following
happens:

-Samus becomes invincible, i.e. neither enemies or enemy bullets or
megaman can kill him

-Invincibility lasts 13 seconds, with audio that builds up to a
crescendo of power, then goes down, letting the player know when the
powerup begins and ends. By design, this audio plays instead of the
normal music, to make it more dramatic.

-If enemies or megaman collide with Samus while Invincible, then they
die, and make collision and dying sounds to further the dramatic effect.